backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-dotenv==1.0.1
httpx==0.27.0
apscheduler==3.10.4
notion-client==2.2.1
python-dateutil==2.9.0
